Update README.md
Task: 
Create a Lambda function to count the number of words in a text file (function code saved in this repo)

Configure an Amazon Simple Storage Service (Amazon S3) bucket to invoke a Lambda function when a text file is uploaded to the S3 bucket. (you so this by creating a S3 bucket then adding S3 as a trigger on lambda

Create an Amazon Simple Notification Service (Amazon SNS) topic to report the word count in an email. You do this by creating a topic first, then adding a subscription. The topic ARN is what you'll use in the Lambda function